Fixed occasional incorrect percentages being printed by /Measure, /Cancel, and /Undo (due to an integer overflow). Updating version numbers for the inevitable 0.619 release.

+++ fCraft/Drawing/DrawOperation.cs
@@ -91,7 +91,7 @@ namespace fCraft.Drawing {
                 }else if( IsDone ) {
                     return 100;
                 } else {
-                    return Math.Min( 100, Math.Max( 0, (BlocksProcessed * 100) / BlocksTotalEstimate ) );
+                    return (int)Math.Min( 100, Math.Max( 0, (BlocksProcessed * 100L) / BlocksTotalEstimate ) );
                 }
             }
         }
